test(cli): consolidate command output coverage
